+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/media-package/compare/0.4.5...develop)
+## [Unreleased](https://github.com/ixocreate/media-package/compare/0.4.6...develop)
+
+## [v0.4.5 (2019-10-11)](https://github.com/ixocreate/media-package/compare/0.4.5...0.4.6)
+### Added
+- ImageAnnotated Element & Type
+### Changed
+- improve memory usage
+- restrict parallel image processing to cli
 
 ## [v0.4.5 (2019-07-18)](https://github.com/ixocreate/media-package/compare/0.4.4...0.4.5)
 ### Changed